requirements.txt
flask
flask-cors
flask-bcrypt
flask-session
pymongo
python-dotenv
requests
ruuvitag-sensor